requirement.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
scapy==2.5.0
SQLAlchemy==2.0.31
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic==2.8.2